requirements.txt
streamlit>=1.29.0
boto3>=1.34.0
pandas>=2.0.0
numpy>=1.26.0
plotly>=5.17.0
python-dotenv>=1.0.0
requests>=2.31.0
scikit-learn>=1.3.0